pyproject.toml
[project]
name = "code-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygments>=2.20.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "textual>=8.2.4",
]
